Fix tests: IE11 incompatible use of `new Set()` (with arguments)
IE11 [doesn't support collection arguments][1].

Disabling the check means that linting will pass, when it should not.
Since the aim is to keep this library compatible with IE11, then the tests
should be able to run in IE11.

Instead of disabling the check for use of invalid use `new Set()`,
replace invalid calls to `new Set()` with a helper `createSet()` that
supports collection arguments.

[1]: https://developer.mozilla.org/en-US/docs/Web/JavaScript/Reference/Global_Objects/Set