Improve isolation of is-element.test.js
During some other work it became clear that this test does not have
good isolation, ad it would fail because `is-element.js` would not
detect a DOM when it was loaded as a side effect of loading `samsam` by
other tests.

The solution is to ensure that this test prepares the environment
properly, by deleting `is-element.js` from `require.cache` and setting
up a DOM, before loading `is-element.js`.

See:
* https://nodejs.org/api/modules.html#modules_require_cache
* http://derpturkey.com/reload-module-with-node-js-require/