Remove use of .callee
IE11 is the only IE supported in Sinon projects.

In IE11, calling toString on an arguments object returns
"[object Arguments]", just like all the modern engines.

Therefore, it is safe to remove the branch that uses `.callee` for
detecting the arguments object.

This helps us get closer to 100% test coverage.