Distribute package as source
This library is not meant for writing end user applications or even for
being used directly when writing tests. It is not meant to be loaded
directly by browsers.

Consumers of this package are assumed to use their own bundlers, should
they need to bundle code for use in browsers or workers.

Therefore, it is safe to distribute the package as source files and not
bundle them up.

This allows us to remove the bundler and its transitive dependencies,
which reduces the maintenance burden of managing the library.