Remove deprecated lodash.get (#254)
We need slightly more than a direct substitution of `.get()` with `?.` because `?.` handles individual properties and hasNested function needs to handle nested paths which are not single properties.

To handle nested paths, the regex is breaking the property string into its individual components and removes empty values.  The loop traverses the object, moving one level deeper for each part of the path so we can safely access the next property (part) of the current object (`current = current?.[part]`).  This handles cases where any intermediate property might be `undefined` or `null`.  If `current` becomes `undefined` at any point we know the path doesn't exist, so we can immediately return `false`.  After the loop is finished then the value of `current` is the value we want to get.